+++ example/SampleApp/www/js/index.js
@@ -69,7 +69,8 @@ var app = {
     locations: [],
     isTracking: false,
     postingEnabled: false,
-    postUrl: 'https://background-geolocation-console.herokuapp.com/locations',
+    // thank you IBM Bluemix
+    postUrl: 'https://bgconsole.mybluemix.net/locations',
     /**
     * @private
     */
@@ -157,6 +158,15 @@ var app = {
         app.receivedEvent('deviceready');
         window.addEventListener('batterystatus', app.onBatteryStatus, false);
         app.configureBackgroundGeoLocation();
+        backgroundGeoLocation.watchLocationMode(app.onLocationCheck);
+    },
+    onLocationCheck: function (enabled) {
+        if (app.isTracking && !enabled) {
+            var showSettings = window.confirm('No location provider enabled. Should I open location setting?');
+            if (showSettings === true) {
+                backgroundGeoLocation.showLocationSettings();
+            }
+        }
     },
     onBatteryStatus: function(ev) {
         app.battery = {
@@ -194,13 +204,10 @@ var app = {
     onOffline: function() {
         console.log('Offline');
     },
-    stop: function () {
-
-    },
     configureBackgroundGeoLocation: function() {
-        var bgGeo = window.plugins.backgroundGeoLocation;
         var anonDevice = {
             model: device.model,
+            version: device.version,
             platform: device.platform,
             uuid: md5([device.uuid, this.salt].join())
         };
@@ -209,7 +216,7 @@ var app = {
         * This would be your own callback for Ajax-requests after POSTing background geolocation to your server.
         */
         var yourAjaxCallback = function(response) {
-            bgGeo.finish();
+            backgroundGeoLocation.finish();
         };
 
         /**
@@ -221,7 +228,8 @@ var app = {
                     uuid: new Date().getTime(),
                     timestamp: location.time,
                     battery: app.battery,
-                    coords: location
+                    coords: location,
+                    service_provider: ENV.settings.locationService
                 },
                 device: anonDevice
             };
@@ -254,7 +262,7 @@ var app = {
         };
 
         // Only ios emits this stationary event
-        bgGeo.onStationary(function(location) {
+        backgroundGeoLocation.onStationary(function(location) {
             if (!app.stationaryRadius) {
                 app.stationaryRadius = new google.maps.Circle({
                     fillColor: '#cc0000',
@@ -271,7 +279,7 @@ var app = {
         });
 
         // BackgroundGeoLocation is highly configurable.
-        bgGeo.configure(callbackFn, failureFn, {
+        backgroundGeoLocation.configure(callbackFn, failureFn, {
             desiredAccuracy: 0,
             stationaryRadius: 50,
             distanceFilter: 50,
@@ -282,18 +290,17 @@ var app = {
             activityType: 'AutomotiveNavigation',
             debug: true, // <-- enable this hear sounds for background-geolocation life-cycle.
             stopOnTerminate: false, // <-- enable this to clear background location settings when the app terminates
-            locationService: bgGeo.service[ENV.settings.locationService]
+            locationService: backgroundGeoLocation.service[ENV.settings.locationService]
         });
 
         // Turn ON the background-geolocation system.  The user will be tracked whenever they suspend the app.
         var settings = ENV.settings;
 
         if (settings.enabled == 'true') {
-            bgGeo.start();
-            app.isTracking = true;
+            app.startTracking();
 
             if (settings.aggressive == 'true') {
-                bgGeo.changePace(true);
+                backgroundGeoLocation.changePace(true);
             }
         }
     },
@@ -325,22 +332,20 @@ var app = {
         }
     },
     onServiceChange: function(ev) {
-        var bgGeo = window.plugins.backgroundGeoLocation,
-            locationService = $(ev.target).text();
+        var locationService = $(ev.target).text();
 
         ENV.settings.locationService = locationService;
         localStorage.setItem('locationService', locationService);
         if (app.isTracking) {
-            bgGeo.stop();
+            app.stopTracking();
             app.configureBackgroundGeoLocation();
-            bgGeo.start();
+            app.startTracking();
         } else {
             app.configureBackgroundGeoLocation();
         }
     },
     onClickChangePace: function(value) {
-        var bgGeo   = window.plugins.backgroundGeoLocation,
-            btnPace = app.btnPace;
+        var btnPace = app.btnPace;
 
         btnPace.removeClass('btn-success');
         btnPace.removeClass('btn-danger');
@@ -348,10 +353,10 @@ var app = {
         var isAggressive = ENV.toggle('aggressive');
         if (isAggressive == 'true') {
             btnPace.addClass('btn-danger');
-            bgGeo.changePace(true);
+            backgroundGeoLocation.changePace(true);
         } else {
             btnPace.addClass('btn-success');
-            bgGeo.changePace(false);
+            backgroundGeoLocation.changePace(false);
         }
     },
     onClickReset: function() {
@@ -369,8 +374,7 @@ var app = {
         }
     },
     onClickToggleEnabled: function(value) {
-        var bgGeo       = window.plugins.backgroundGeoLocation,
-            btnEnabled  = app.btnEnabled,
+        var btnEnabled  = app.btnEnabled,
             isEnabled   = ENV.toggle('enabled');
 
         btnEnabled.removeClass('btn-danger');
@@ -379,13 +383,11 @@ var app = {
         if (isEnabled == 'true') {
             btnEnabled.addClass('btn-danger');
             btnEnabled[0].innerHTML = 'Stop';
-            bgGeo.start();
-            app.isTracking = true;
+            app.startTracking();
         } else {
             btnEnabled.addClass('btn-success');
             btnEnabled[0].innerHTML = 'Start';
-            bgGeo.stop();
-            app.isTracking = false;
+            app.stopTracking();
         }
     },
     /**
@@ -406,6 +408,15 @@ var app = {
     // Update DOM on a Received Event
     receivedEvent: function(id) {
         console.log('Received Event: ' + id);
+    },
+    startTracking: function () {
+        backgroundGeoLocation.start();
+        app.isTracking = true;
+        backgroundGeoLocation.isLocationEnabled(app.onLocationCheck);
+    },
+    stopTracking: function () {
+        backgroundGeoLocation.stop();
+        app.isTracking = false;
     },
     setCurrentLocation: function(location) {
         var map = app.map;